docker/aws_code/fastapi2.9.0/event/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
mysql-connector-python
python-multipart
boto3
requests
pandas
PyJWT
cryptography
